+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
 ## dev-develop
+
+ - ENHANCEMENT #24    Update sulu from `alexander-schranz/sulu-form-bundle` from commit `c110e72e44a58a0f53428c153e405124695506f6`
+ - BUGFIX      #24    Fixed form preview request analyzer
+ - BUGFIX      #24    Fixed second date field type
+ - BUGFIX      #24    Fixed missing translation and documentation
+ - ENHANCEMENT #24    Added missing documentation
+ - BUGFIX      #24    Fixed success email not sent to email in data json
+ - BUGFIX      #24    Fixed missing uniqueness in form fields table
+ - BUGFIX      #24    Fixed permission denied on fields action
+ - FEATURE     #24    Added toggler to set email as replyTo
+ - BUGFIX      #24    Fixed exception controller redirect
+ - BUGFIX      #24    Fixed form select by using native select
+ - FEATURE     #24    Added media collection strategy tree
+ - FEATURE     #24    Added support to add dynamic list to article bundle
+ - BUGFIX      #24    Fixed csv export for bool values
+ - ENHANCEMENT #24    Fixed sort on dynamic list and removed search
+ - FEATURE     #24    Added dynamic form list and export
  - FEATURE     #20    Added additional receivers for sending notification as CC or BCC
  - ENHANCEMENT #23    Fixed composer json links and description
  - BUGFIX      #16    Fixed setting of entity value in dynamic which represents an array
@@ -15,4 +32,3 @@
  - FEATURE     #2     Changed title field for labels to texteditor
  - ENHANCEMENT #-     Updated namespaces and rename table names
  - ENHANCEMENT #-     Forked from `alexander-schranz/sulu-form-bundle` from commit `36a7cd11562ed0c9f64752b37707cc2771e0baca`
-